refactor: 폴더 구조 이동
