+++ src/main/java/com/ll/commars/domain/auth/entity/AuthResponse.java
@@ -1,4 +1,4 @@
-package com.ll.commars.domain.auth.auth.entity;
+package com.ll.commars.domain.auth.entity;
 
 import lombok.Getter;
 import lombok.Setter;
