+++ src/main/java/com/ll/commars/domain/auth/google/controller/ApiV1AuthGoogleController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.ll.commars.domain.auth.auth.service.AuthService;
+import com.ll.commars.domain.auth.service.AuthService;
 import com.ll.commars.domain.auth.google.service.GoogleService;
 import com.ll.commars.domain.user.entity.User;
