+++ src/main/java/com/ll/commars/domain/auth/kakao/controller/ApiV1AuthKakaoController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.ll.commars.domain.auth.auth.service.AuthService;
+import com.ll.commars.domain.auth.service.AuthService;
 import com.ll.commars.domain.auth.kakao.service.KakaoService;
 import com.ll.commars.domain.user.entity.User;
