+++ src/main/java/com/ll/commars/domain/auth/service/AuthService.java
@@ -1,11 +1,11 @@
-package com.ll.commars.domain.auth.auth.service;
+package com.ll.commars.domain.auth.service;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import com.ll.commars.domain.auth.auth.entity.AuthResponse;
+import com.ll.commars.domain.auth.entity.AuthResponse;
 import com.ll.commars.domain.user.entity.User;
 import com.ll.commars.global.jwt.component.JwtProvider;
